requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asyncio==4.0.0
attrs==25.3.0
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
frozenlist==1.7.0
h11==0.16.0
idna==3.10
lxml==6.0.1
magic-filter==1.0.12
multidict==6.6.4
numpy==2.3.2
outcome==1.3.0.post0
pandas==2.3.2
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
